feat: Added Starters for Jdbc, Hazelcast, Redis, Mongodb backends
