+++ services/aggregator/pom.xml
@@ -9,26 +9,6 @@
   <name>Activiti Cloud Services :: Messages Aggregator</name>
   <artifactId>activiti-cloud-services-messages-aggregator</artifactId>
   
-  <properties>
-    <spring-integration-hazelcast.version>2.0.0.RELEASE</spring-integration-hazelcast.version>
-    <hazelcast.version>3.12.4</hazelcast.version>
-  </properties>
-  
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>com.hazelcast</groupId>
-        <artifactId>hazelcast</artifactId>
-        <version>${hazelcast.version}</version>
-      </dependency>  
-      <dependency>
-        <groupId>com.hazelcast</groupId>
-        <artifactId>hazelcast-spring</artifactId>
-        <version>${hazelcast.version}</version>
-      </dependency>  
-    </dependencies>
-  </dependencyManagement>
-  
   <dependencies>
     <!-- Activiti Cloud -->          
     <dependency>
@@ -40,10 +20,6 @@
       <groupId>com.fasterxml.jackson.datatype</groupId>
       <artifactId>jackson-datatype-jdk8</artifactId>
     </dependency>
-    <dependency>
-      <groupId>com.hazelcast</groupId>
-      <artifactId>hazelcast-spring</artifactId>
-    </dependency>  
     <!-- Spring -->
     <dependency>
       <groupId>org.springframework</groupId>
@@ -69,62 +45,14 @@
     <!-- Spring Integration -->
     <dependency>
       <groupId>org.springframework.integration</groupId>
-      <artifactId>spring-integration-hazelcast</artifactId>
-      <version>${spring-integration-hazelcast.version}</version>      
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.integration</groupId>
       <artifactId>spring-integration-core</artifactId>
     </dependency>  
-    <dependency>
-      <groupId>org.springframework.integration</groupId>
-      <artifactId>spring-integration-jdbc</artifactId>    
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.integration</groupId>
-      <artifactId>spring-integration-redis</artifactId>    
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.integration</groupId>
-      <artifactId>spring-integration-mongodb</artifactId>    
-    </dependency>  
-    <!-- Runtime -->
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-data-redis</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-data-mongodb</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-jdbc</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.hsqldb</groupId>
-      <artifactId>hsqldb</artifactId>
-      <scope>runtime</scope>
-    </dependency>
+    <!-- Test -->
     <dependency>
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
-      <scope>runtime</scope>
-    </dependency>    
-    <dependency>
-      <groupId>org.mariadb.jdbc</groupId>
-      <artifactId>mariadb-java-client</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.postgresql</groupId>
-      <artifactId>postgresql</artifactId>
-      <scope>runtime</scope>
-    </dependency>    
-    <!-- Test -->
+      <scope>test</scope>
+    </dependency>        
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
