+++ services/aggregator/src/main/java/org/activiti/cloud/services/messages/aggregator/config/MessageConnectorIntegrationConfiguration.java
@@ -54,6 +54,7 @@ import org.springframework.integration.aggregator.MessageGroupProcessor;
 import org.springframework.integration.aggregator.ReleaseStrategy;
 import org.springframework.integration.config.EnableIntegration;
 import org.springframework.integration.config.EnableIntegrationManagement;
+import org.springframework.integration.config.EnableMessageHistory;
 import org.springframework.integration.dsl.IntegrationFlow;
 import org.springframework.integration.dsl.IntegrationFlows;
 import org.springframework.integration.dsl.MessageChannels;
@@ -73,7 +74,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @Configuration
 @EnableIntegration
 @EnableBinding(MessageConnectorProcessor.class)
-//@EnableMessageHistory // FIXME does not work with RabbitMQ binder
+@EnableMessageHistory 
 @EnableIntegrationManagement
 @EnableConfigurationProperties(MessageAggregatorProperties.class)
 @EnableTransactionManagement
