+++ services/aggregator/src/main/resources/META-INF/spring.factories
@@ -1,7 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.activiti.cloud.services.messages.aggregator.config.MessageStoreAutoConfiguration,\
     org.activiti.cloud.services.messages.aggregator.config.MessageConnectorAutoConfiguration
-    
-org.springframework.boot.env.EnvironmentPostProcessor=\
-  org.activiti.cloud.services.messages.aggregator.config.MessageStoreEnvironmentPostProcessor
