--- services/aggregator/src/test/java/org/activiti/cloud/services/messages/aggregator/test/SimpleMessageStoreTests.java
@@ -1,17 +1,0 @@
-package org.activiti.cloud.services.messages.aggregator.test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.Test;
-import org.springframework.integration.store.SimpleMessageStore;
-import org.springframework.test.context.TestPropertySource;
-
-@TestPropertySource(properties = {
-        "activiti.cloud.services.messages.aggregator.message-store-type=simple"})
-public class SimpleMessageStoreTests extends AbstractIntegrationFlowTests {
-
-    @Test
-    public void testMessageStore() throws Exception {
-        assertThat(this.aggregatingMessageHandler.getMessageStore()).isInstanceOf(SimpleMessageStore.class);
-    }
-}
