+++ starters/redis/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+    org.activiti.cloud.starter.messages.redis.RedisMessageStoreAutoConfiguration
+
